requirements.txt
django==1.11.6
pillow==5.0.0
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
markdown2==2.3.5
pdfkit==0.6.1
django-easy-pdf>=0.2.0 
WeasyPrint>=0.34